+++ src/EntityGraphQL/Schema/BaseSchemaTypeWithFields.cs
@@ -110,7 +110,7 @@ namespace EntityGraphQL.Schema
             FieldsByName.Remove(name);
         }
 
-        public abstract ISchemaType AddAllBaseTypes();
+        public abstract ISchemaType ImplementAllBaseTypes(bool addTypeIfNotInSchema = true, bool addAllFieldsOnAddedType = true);
         public abstract ISchemaType Implements<TClrType>(bool addTypeIfNotInSchema = true, bool addAllFieldsOnAddedType = true);
         public abstract ISchemaType Implements(string typeName);
     }